fix: add tool to thread state as a string
Since the tool parameter is passed here as a string and the former
for-loop used tool as a Tool object, this caused the tool that was saved
in the thread state to be an object rather than a string. This would
cause errors when reloading the thread from the file.

Signed-off-by: Donnie Adams <donnie@acorn.io>