fix: stop changing icon during chat
Before this change, the chat icon would change based on the tool that
was being called. For example, if an assistant was using a tool that
needed to prompt the user, then the icon would change to the initials
for that tool. After this change, the icon will only change when the
assistant changes.

A side effect that is also addressed in this change: restoring a thread
now has the correct icons for the messages. This will only be true for
threads created after this change.

Signed-off-by: Donnie Adams <donnie@acorn.io>