feat: add ability to save an assistant
If a user has added tools to a thread, then they will be able to save
the assistant such that these tools become part of the assistant. The
only exception to this is the knowledge tool: the user can't explicitly
add the knowledge tool, so it will not be saved to the assistant. This
will change when a user has the ability to add knowledge to an
assistant.

Signed-off-by: Donnie Adams <donnie@acorn.io>