fix: clear tools when new thread is created
When switching threads, the socket will send back the messages and tools
for the thread. However, if the thread state file doesn't exist yet,
nothing is sent. This leads to the tools being leftover from the
previous thread.

This change will emit the event so that the tools are correctly set for
a new thread.

Signed-off-by: Donnie Adams <donnie@acorn.io>