fix: try parsing tools added by URL to detect errors
If an error is detected when parsing a tool added by URL, then the error
is displayed to the user and the tool is not added. The user will have
the opportunity to adjust the URL and try again.

Additionally, while the parsing is happening, a spinner has been added
to indicate that something is happening.

Signed-off-by: Donnie Adams <donnie@acorn.io>