fix: ensure that runs are stopped when switch threads
This change will ensure that currently running gptscript runs are
stopped when switching threads. Additionally, this change removes the
redundant selectedThreadId state because it is the same as thread.

Finally, we stop displaying the "run aborted" errors to the user because
it isn't really an error: the user has chosen to abort the run.

Signed-off-by: Donnie Adams <donnie@acorn.io>