fix: fetch threads after using "Save As" on a thread
When a user uses the "Save Assistant As" functionality, the new
assistant they are creating is supposed to replace the current assistant
in the thread. This was happening correctly.

However, the threads were being stored in state. Therefore, navigating
away from the thread and coming back to it would make it look like the
old assistant was still being used. This change will refresh the thread
state after "Save Assistant As" so that the state will reflect the new
assistant.

Signed-off-by: Donnie Adams <donnie@acorn.io>