fix: improve multiple prompts (#463)
* fix: improve multiple prompts

When a prompt was followed by another prompt, React was "helping" us by
not completing resetting the state. This caused an issue where the old
state would be used in the next prompt, making it appear as though the
prompt was already submitted.

This change will reset this state so it will not be reused.

This change also includes some small changes around what is displayed
before and after prompts. For example, when an authentication prompt
disappeared, the message would still say,
"<Tool> needs to authenticate ." Now, the message will switch back to
"Waiting for model response..."

Signed-off-by: Donnie Adams <donnie@acorn.io>