fix: stop error when tool has no instructions
Previously, if a tool had no instructions, then the SDK would report an
error because there would have been no output. That has been changed in
the SDK.

Now, in these changes, if a blank message is received when running a
tool, a special "no response received" message will be displayed to the
user.

Signed-off-by: Donnie Adams <donnie@acorn.io>